Changed file struct
