+++ backend/routes/Login.js
@@ -1,86 +1,68 @@
-// const { body, check, validationResult } = require("express-validator");
-// const {database} = require("../database/database")
-// const jwt = require('jsonwebtoken')
-// const dotenv = require("dotenv");
+const { body, check, validationResult } = require("express-validator");
+const {database} = require("../database/database")
+const jwt = require('jsonwebtoken')
+const dotenv = require("dotenv");
 
-// dotenv.config();
+dotenv.config();
 
 
-// const registerUser = (req, res, next) => 
-// {
-
-//     const errors = validationResult(req);
-
-//     if (!errors.isEmpty())
-//     {
-//         return res.status(400).json({ errors: errors.array() });
-//     }
-
-//     //console.log(req.body)
-//     let data = {
-//         username: req.body.username,
-//         password: req.body.password,
-//         history: [], //an array of json objects
-//         fullName: "",
-//         company: "",
-//         address1: "",
-//         address2: "",
-//         city: "",
-//         zipcode: "",
-//         state: "",
-//         id: database.length + 1,
-//         authToken: "",
-//       };
+const registerUser = (req, res, next) => 
+{
+    console.log(req.body);
+    let data = {
+        username: req.body.username,
+        password: req.body.password,
+        history: [], //an array of json objects
+        fullName: "",
+        company: "",
+        address1: "",
+        address2: "",
+        city: "",
+        zipcode: "",
+        state: "",
+        id: database.length + 1,
+        token: "",
+      };
     
-//       database.push(data)
-//       res.json(database);
+      database.push(data)
+      //console.log(database)
+      res.json(database);
 
-// }
+}
 
 
-// const logUserIn = (req, res, nex) => 
-// {
+const logUserIn = (req, res, nex) => 
+{
 
-//     // const errors = validationResult(req);
+//     console.log(req.body);
+  const userData = database.find(
+    (user) => user.username === req.body.username
+  );
+  let pass;
+  if (userData) {
+    // console.log(userData)
+    if (userData.password === req.body.password){
+      req.session.user = userData
+      //console.log(req.session.user)
+      pass = userData.password;
 
-//     // if (!errors.isEmpty())
-//     // {
-//     //     return res.status(400).json({ errors: errors.array() });
-//     // }
-    
-//     console.log(req.body)
-//     const username = req.body.username;
-//     const password = req.body.password;
+      //Auth
+      const id = userData.id
+      
+      const token = jwt.sign({id}, process.env.ACCESS_TOKEN_SECRET);
 
-//     const userData = database.find(
-//         (user) => user.username === req.body.username
-//       );
-//       let pass;
-//       if (userData) {
-//         //console.log(userData)
-//         if (userData.password == req.body.password){
-          
-//           //console.log(req.session.user)
-//           pass = userData.password;
-    
-//           //Auth
-//           const id = userData.id
-//           const token = jwt.sign({id}, process.env.ACCESS_TOKEN_SECRET, {
-//             expiresIn: 300,
-//           });
+      userData.token = token; 
 
-//           req.session.user = userData
+      res.json({auth: true, token: token, userData: userData})
+    }
+    else {
+      res.json({auth: false, message: "Not Found Pass!"})
+    }
+  } else {
+    res.json({auth: false, message: "Not Found User!"})
+  }
 
-//           res.json({authentication: true, token: token, userId: userData.id})
-//         }
-//         else {
-//             return res.json({authentication: false, message: "Password doesn't match"})
-//         }
-//       } else {
-//         return res.json({authentication: false, message: "User not found"})
-//       }
-
-// }
+}
 
 
 // const checkAuth = (res, req, nex) =>
@@ -107,9 +89,9 @@
 
 
 
-// module.exports = 
-// {
-//     registerUser,
-//     logUserIn,
+module.exports = 
+{
+     registerUser,
+     logUserIn,
 //     checkAuth,
-// }
+}
